requirements.txt
streamlit
rdkit
pandas
numpy
chembl_webresource_client.new_client